+++ src/collective/alerts/browser/alert_viewlet.py
@@ -15,9 +15,9 @@ class AlertViewlet(ViewletBase):
         super(AlertViewlet, self).update()
         alerts = getattr(self.context, 'enable_alerts', False)
         if alerts:
-            self.show_in_context = "true"
+            self.show_in_context = True
         else:
-            self.show_in_context = "false"
+            self.show_in_context = False
         self.get_message_view = "%s/get-alert-message" % self.context.portal_url.getPortalObject().absolute_url()
         self.get_global_message_view = (
                 "%s/get-global-alert-message" %
@@ -42,6 +42,7 @@ class AlertViewlet(ViewletBase):
 
     def json_config(self, kind="local"):
         results = dict()
+        results["show_in_context"] = self.show_in_context
         if kind == "local":
             results["get_message_view"] = self.get_message_view
             results["cache"] = self.cache_view