+++ CHANGES.rst
@@ -2,7 +2,7 @@ Changelog
 =========
 
 
-2.2.1 (unreleased)
+2.3.0 (2019-12-10)
 ------------------
 
 - Allow to modify the alert types from the control panel